write hook for control navbar style.
